Update en_US.ini
If "FPS" are not in capital the string won't get translated. Also I
found a bug that it won't let "FPS" string translated on whatever I'll
change it (FPS, FPS: FPS   :, etc). It may need to be changed into
something else on the Qt since the translation mechanism doesn't like
strings with a ":" on them.